Python-Projects/Django-API/services/pyproject.toml
[tool.poetry]
name = "services"
version = "0.1.0"
description = ""
authors = ["kkf017 <cpunk1017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
